requirements.txt
imaplib2
python-dotenv
tqdm